refac) container component 생성
