+++ src/InputContainer.jsx
@@ -1,0 +1,31 @@
+import React from 'react';
+
+import { useDispatch, useSelector } from 'react-redux';
+
+import { updateTaskTitle, addTask } from './actions';
+
+import Input from './Input';
+
+export default function App() {
+  const dispatch = useDispatch();
+
+  const { taskTitle } = useSelector((state) => {
+    taskTitle: state.taskTitle,
+  });
+
+  function handleChangeTitle(event) {
+    dispatch(updateTaskTitle(event.target.value));
+  }
+
+  function handleClickAddTask() {
+    dispatch(addTask());
+  }
+
+  return (
+    <Input
+      taskTitle={taskTitle}
+      onChangeTitle={handleChangeTitle}
+      onClickAddTask={handleClickAddTask}
+    />
+  );
+}
